requirements.txt
# --- Core Backend Framework ---
fastapi==0.110.0
uvicorn[standard]==0.29.0

# --- PDF Processing ---
PyPDF2==3.0.1
pdfminer.six==20221105

# --- Machine Learning / Embeddings ---
sentence-transformers>=2.7.0       # upgraded for HuggingFace compatibility
torch>=2.0.0
transformers>=4.40.0               # ensure latest features
huggingface-hub>=0.22.0            # needed for hf_hub_download API

# --- Vector Database (Qdrant) ---
qdrant-client==1.7.3

# --- Database ORM (for metadata & query history) ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

# --- Environment Variables ---
python-dotenv==1.0.1

# --- Data Handling ---
pandas==2.2.2
numpy==1.26.4

# --- Testing ---
pytest==8.2.0
requests==2.31.0

# --- Optional ---
httpx==0.27.0  # For async API calls if needed

# --- File Upload Support ---
python-multipart>=0.0.6